ScalaSigBytes is no longer a ClassfileAnnotArg
This abstraction was not necessary, it was just dealing with encoding
the signature. This is now handled more directly.
