+++ src/reflect/scala/reflect/internal/AnnotationInfos.scala
@@ -112,55 +112,6 @@ trait AnnotationInfos extends api.Annotations { self: SymbolTable =>
   val NestedArgument = NestedAnnotArg
   implicit val NestedArgumentTag = ClassTag[NestedAnnotArg](classOf[NestedAnnotArg])
 
-  /** A specific annotation argument that encodes an array of bytes as an
-   *  array of `Long`. The type of the argument declared in the annotation
-   *  must be `String`. This specialised class is used to encode Scala
-   *  signatures for reasons of efficiency, both in term of class-file size
-   *  and in term of compiler performance.
-   *  Details about the storage format of pickles at the bytecode level (classfile annotations) can be found in SIP-10.
-   */
-  case class ScalaSigBytes(bytes: Array[Byte]) extends ClassfileAnnotArg {
-    override def toString = (bytes map { byte => (byte & 0xff).toHexString }).mkString("[ ", " ", " ]")
-    lazy val sevenBitsMayBeZero: Array[Byte] = {
-      mapToNextModSevenBits(scala.reflect.internal.pickling.ByteCodecs.encode8to7(bytes))
-    }
-
-    /* In order to store a byte array (the pickle) using a bytecode-level annotation,
-     * the most compact representation is used (which happens to be string-constant and not byte array as one would expect).
-     * However, a String constant in a classfile annotation is limited to a maximum of 65535 characters.
-     * Method `fitsInOneString` tells us whether the pickle can be held by a single classfile-annotation of string-type.
-     * Otherwise an array of strings will be used.
-     */
-    def fitsInOneString: Boolean = {
-      // due to escaping, a zero byte in a classfile-annotation of string-type takes actually two characters.
-      var i = 0
-      var numZeros = 0
-      while (i < sevenBitsMayBeZero.length) {
-        if (sevenBitsMayBeZero(i) == 0) numZeros += 1
-        i += 1
-      }
-
-      (sevenBitsMayBeZero.length + numZeros) <= 65535
-    }
-
-    def sigAnnot: Type =
-      if (fitsInOneString)
-        definitions.ScalaSignatureAnnotation.tpe
-      else
-        definitions.ScalaLongSignatureAnnotation.tpe
-
-    private def mapToNextModSevenBits(src: Array[Byte]): Array[Byte] = {
-      var i = 0
-      val srclen = src.length
-      while (i < srclen) {
-        val in = src(i)
-        src(i) = (if (in == 0x7f) 0.toByte else (in + 1).toByte)
-        i += 1
-      }
-      src
-    }
-  }
-
   object AnnotationInfo {
     def marker(atp: Type): AnnotationInfo =
       apply(atp, Nil, Nil)
